requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.8
pydantic-settings>=2.4
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
joblib>=1.4
sqlalchemy>=2.0
pyodbc>=5.1
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
httpx>=0.27
structlog>=24.1
prometheus-client>=0.20
python-multipart>=0.0.9
huggingface_hub>=0.24
python-dotenv
openai 
pydantic